Split SQLite schema alignment tests
Convert the oversized SQLite schema-alignment test file into focused
submodules for shared fixtures, catalogue checks, GUID/counter behaviour,
upgrade flows, and article threading.

Update the developers' guide so the documented test harness layout matches
the new directory module structure.
